make downloadReferences.sh resumable

+++ installation/downloadReferences.sh
@@ -19,16 +19,16 @@ mkdir -p hg19_GRCh37_1000genomes &&
 cd hg19_GRCh37_1000genomes ||
 exit_on_fail
 
-if [[ "$REFERENCE_GENOME" == "true" ]] 
+if [[ "$REFERENCE_GENOME" == "true" && ! -e sequence/1KGRef/hs37d5.fa ]] 
 then
 	echo downloading reference genome....
 	mkdir -p sequence/1KGRef &&
-	wget -P sequence/1KGRef ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/technical/reference/phase2_reference_assembly_sequence/hs37d5.fa.gz &&
+	wget -c -P sequence/1KGRef ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/technical/reference/phase2_reference_assembly_sequence/hs37d5.fa.gz &&
 	gunzip sequence/1KGRef/hs37d5.fa.gz ||
 	exit_on_fail
 fi
 
-if [[ "$dbSNP_FILE" == "true" ]]
+if [[ "$dbSNP_FILE" == "true" && ! -e databases/dbSNP/dbSNP_135/00-All.SNV.vcf.gz.tbi ]]
 then
 	echo downloading dbSNP file....
 	mkdir -p databases/dbSNP/dbSNP_135 &&
@@ -58,11 +58,11 @@ then
 	exit_on_fail
 fi
 
-if [[ "$MAPPABILITY_FILE" == "true" ]]
+if [[ "$MAPPABILITY_FILE" == "true" && ! -e databases/UCSC/wgEncodeCrgMapabilityAlign100mer_chr.bedGraph.gz.tbi ]]
 then
 	echo downloading mappability file....
 	mkdir -p databases/UCSC &&
-	wget -P databases/UCSC http://hgdownload.soe.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeMapability/wgEncodeCrgMapabilityAlign100mer.bigWig &&
+	wget -c -P databases/UCSC http://hgdownload.soe.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeMapability/wgEncodeCrgMapabilityAlign100mer.bigWig &&
 	bigWigToBedGraph databases/UCSC/wgEncodeCrgMapabilityAlign100mer.bigWig databases/UCSC/wgEncodeCrgMapabilityAlign100mer_chr.bedGraph &&
 	rm -f databases/UCSC/wgEncodeCrgMapabilityAlign100mer.bigWig &&
 	bgzip databases/UCSC/wgEncodeCrgMapabilityAlign100mer_chr.bedGraph &&
@@ -70,7 +70,7 @@ then
 	exit_on_fail
 fi
 
-if [[ "$CHROMOSOME_LENGTH_FILE" == "true" ]] 
+if [[ "$CHROMOSOME_LENGTH_FILE" == "true" && ! -e stats/chrlengths.txt ]] 
 then
 	echo downloading dbSNP file....
 	mkdir -p stats &&
@@ -78,7 +78,7 @@ then
 	exit_on_fail
 fi
 
-if [[ "$statFiles" == "true" ]]
+if [[ "$statFiles" == "true" && ! -e databases/ENCODE/ReplicationTime_10cellines_mean_10KB.Rda ]]
 then
 	echo downloading statsfile....
 	mkdir -p stats &&
@@ -88,14 +88,14 @@ then
 	exit_on_fail
 fi
 
-if [[ "$IMPUTE_FILES" == "true" ]]
+if [[ "$IMPUTE_FILES" == "true" && ! -e databases/1000genomes/IMPUTE/ALL_1000G_phase1integrated_v3_impute/ ]]
 then
 	echo downloading impute files....
 	mkdir -p databases/1000genomes/IMPUTE &&
-	wget -P databases/1000genomes/IMPUTE https://mathgen.stats.ox.ac.uk/impute/ALL.integrated_phase1_SHAPEIT_16-06-14.nomono.tgz &&
+	wget -c -P databases/1000genomes/IMPUTE https://mathgen.stats.ox.ac.uk/impute/ALL.integrated_phase1_SHAPEIT_16-06-14.nomono.tgz &&
 	tar -xzvf databases/1000genomes/IMPUTE/ALL.integrated_phase1_SHAPEIT_16-06-14.nomono.tgz -C databases/1000genomes/IMPUTE &&
 	rm -f databases/1000genomes/IMPUTE/ALL.integrated_phase1_SHAPEIT_16-06-14.nomono.tgz &&
-	wget -P databases/1000genomes/IMPUTE https://mathgen.stats.ox.ac.uk/impute/ALL_1000G_phase1integrated_v3_impute.tgz &&
+	wget -c -P databases/1000genomes/IMPUTE https://mathgen.stats.ox.ac.uk/impute/ALL_1000G_phase1integrated_v3_impute.tgz &&
 	tar -xzvf databases/1000genomes/IMPUTE/ALL_1000G_phase1integrated_v3_impute.tgz -C databases/1000genomes/IMPUTE &&
 	rm -f databases/1000genomes/IMPUTE/ALL_1000G_phase1integrated_v3_impute.tgz ||
 	exit_on_fail
